Server now acts as a webserver
Server is now able to handle GET requests and return index.html if
the file path has been correctly given. Still has issues with
error handling, such as if no requests or invalid request is
given. If is; however, able to return an ERROR 404 if the correct
GET request is made but page is not found.