Implement command-line utility functinoalities
Added CLI utilities similar to what Django has. Using python3 manager.py,
on the command-line interface, the program is able to fetch arguments with
sys.argv.

Added functionality to auto-restart the server upon saving a change
to the main file, which is server.py at this moment in time. Currently
it's bugged and will not be restarted due to the recv method, and
if the server is stopped using KeyboardInterrupt, manage.py will rerun
in an infinite loop.